Define AI tools
