+++ packages/core/src/interfaces/ai/agent.ts
@@ -1,0 +1,12 @@
+import { AppContext } from "src/interfaces/app";
+import { Tool } from "./tool";
+
+type RespondInput = {
+  ctx: AppContext;
+  message: string;
+  tools: Tool<any>[];
+};
+
+export type Agent = {
+  respond: (input: RespondInput) => Promise<void>;
+};
